feat: improve post service with validation and error handling for ID operations

+++ src/post/post.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Post } from './schemas/post.schema';
 import mongoose from 'mongoose';
@@ -10,7 +14,9 @@ export class PostService {
   ) {}
 
   async findAll(): Promise<Post[]> {
-    const posts = await this.PostModel.find();
+    const posts = await this.PostModel.find()
+      .populate('author', 'name email')
+      .exec();
     return posts;
   }
 
@@ -20,19 +26,44 @@ export class PostService {
   }
 
   async findById(id: string): Promise<Post | null> {
-    const post = await this.PostModel.findById(id);
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException(`Invalid Post ID: ${id}`);
+    }
+    const post = await this.PostModel.findById(id)
+      .populate('author', 'name email')
+      .exec();
+
+    if (!post) {
+      throw new NotFoundException(`No Post found with ID: ${id}`);
+    }
     return post;
   }
 
   async updateById(id: string, updatedPost: Post): Promise<Post | null> {
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException(`Invalid Post ID: ${id}`);
+    }
     const post = await this.PostModel.findByIdAndUpdate(id, updatedPost, {
       new: true,
       runValidators: true,
     });
+
+    if (!post) {
+      throw new NotFoundException(`No Post found with ID: ${id}`);
+    }
     return post;
   }
 
   async deleteById(id: string): Promise<Post | null> {
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException(`Invalid Post ID: ${id}`);
+    }
     return await this.PostModel.findByIdAndDelete(id);
   }
 }
